+++ test/fixtures/cookbooks/test/recipes/default.rb
@@ -1,58 +1,25 @@
+# Master default recipe that includes all test recipes in logical order
+
+# Disable epel-next for CentOS Stream 10 - repository doesn't exist yet
+# https://github.com/sous-chefs/yum-epel/issues
+if platform?('centos') && node['platform_version'].to_i >= 10
+  node.default['yum']['epel-next']['managed'] = false
+  node.default['yum']['epel-next']['enabled'] = false
+end
+
+apt_update
+
+apt_package 'sudo' if platform_family?('debian')
 gpg_install
 
-%w(foo bar foobar barfoo).each do |u|
-  group u
-  user u do
-    manage_home true
-    gid u
-  end
-end
+# Test 1: Generate GPG keys with various options
+include_recipe 'test::test_generate'
 
-gpg_key 'foo' do
-  user 'foo'
-  passphrase 'this-is-not-secure'
-end
+# Test 2: Export keys to files
+include_recipe 'test::test_export'
 
-# adding bar to root
-gpg_key 'bar' do
-  passphrase 'this-is-not-secure'
-end
+# Test 3: Import keys from files and keyservers
+include_recipe 'test::test_import'
 
-gpg_key 'foobar' do
-  user 'foobar'
-  # override_default_keyring, [true,false], default: false
-  # pubring_file, String
-  # secring_file, String
-  name_real 'Foo Bar'
-  name_comment 'custom comment by foobar'
-  name_email 'foobar@sous-chefs.org'
-  expire_date '20200815T145012'
-  batch_config_file '/tmp/foobar_config'
-  key_type '1'
-  key_length '4096'
-  passphrase 'this-is-not-secure'
-end
-
-gpg_key 'export foo' do
-  user 'foo'
-  name_real 'foo'
-  key_file '/tmp/foo.key'
-  action :export
-end
-
-gpg_key 'import key foo to root keychain' do
-  user 'root'
-  # name_real 'foo-imported'
-  key_file '/tmp/foo.key'
-  action :import
-end
-
-gpg_key 'import key foo to barfoo keychain' do
-  user 'barfoo'
-  # name_real 'foo-imported'
-  key_file '/tmp/foo.key'
-  action :import
-end
-
-# Dummy key for deleting
-include_recipe 'test::dummy_key'
+# Test 4: Delete keys (both public and secret keys)
+include_recipe 'test::test_delete'