+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This file is used to list changes made in each version of the gpg cookbook.
 
 ## [2.0.16](https://github.com/sous-chefs/gpg/compare/2.0.15...v2.0.16) (2025-10-15)
 
+* Add support for adding keys via --keyserver and --recv-keys
+* Remove broken test for user public key
+  * There doesn't seem to be a sensible way to set the home-directory or load the keyring for the user in the test environment
+* Fix installation for Amazon Linux 2023
+* Fix installation on SUSE (add missing package dependencies)
+* Breaking change: only allow 1 key to be imported at a time
+
+## 2.0.14 - *2024-12-05*
 
 ### Bug Fixes